Added README, moved and edit config, added to config.
Added a basic README

Moved the client file out of the client folder

Added better comments to the config.

Added option to Enable/Disable on screen text, as well as watermark.

Cleaned up the code in the config file.

Fixed the use of the wrong config var in the cruise control loop.

Fixed an issue where clicking the cruise control button would not disable the cruise control.

Properly declared all the undeclared local variables.

Replaced `GetPlayerPed(-1)` with `PlayerPedId()`.

Tested In-Game, all works.